+++ Editor/Drawers/SaintsDictionary/SaintsDictionaryDrawerIMGUI.cs
@@ -716,7 +716,7 @@ namespace SaintsField.Editor.Drawers.SaintsDictionary
 
             if (!string.IsNullOrEmpty(cachedInfo.KeySearch) || !string.IsNullOrEmpty(cachedInfo.ValueSearch))
             {
-                refreshedHitTargetIndexes = Search(cachedInfo.KeysProp, cachedInfo.ValuesProp, cachedInfo.KeySearch, cachedInfo.ValueSearch).ToList();
+                refreshedHitTargetIndexes = Search(cachedInfo.KeysProp, cachedInfo.ValuesProp, cachedInfo.KeySearch, cachedInfo.ValueSearch).Where(each => each != -1).ToList();
             }
 
             if (cachedInfo.NumberOfItemsPerPage > 0)