+++ Editor/Utils/SerializedUtils.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using SaintsField.Editor.Drawers.SaintsRowDrawer;
 using SaintsField.Playa;
+using SaintsField.Utils;
 using UnityEditor;
 using UnityEngine;
 using Object = UnityEngine.Object;
@@ -389,23 +390,28 @@ namespace SaintsField.Editor.Utils
             switch (propertyType)
             {
                 case SerializedPropertyType.Integer:
-                    return property.intValue.ToString().Contains(token);
+                    return property.intValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Boolean:
-                    return property.boolValue.ToString().Contains(token);
+                    return property.boolValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Float:
                     // ReSharper disable once SpecifyACultureInStringConversionExplicitly
-                    return property.floatValue.ToString().Contains(token);
+                    return property.floatValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.String:
-                    return property.stringValue?.Contains(token) ?? false;
+                    return property.stringValue?.ToLower().Contains(token) ?? false;
                 case SerializedPropertyType.Color:
-                    return property.colorValue.ToString().Contains(token);
+                    return property.colorValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.ObjectReference:
+                    if (RuntimeUtil.IsNull(property.objectReferenceValue))
+                    {
+                        return false;
+                    }
+
                     // ReSharper disable once Unity.NoNullPropagation
                     if (property.objectReferenceValue is ScriptableObject so)
                     {
                         return SearchSoProp(so, token);
                     }
-                    return property.objectReferenceValue?.name.Contains(token) ?? false;
+                    return property.objectReferenceValue.name.ToLower().Contains(token);
                 case SerializedPropertyType.LayerMask:
                     return property.intValue.ToString().Contains(token);
                 case SerializedPropertyType.Enum:
@@ -414,42 +420,42 @@ namespace SaintsField.Editor.Utils
                     {
                         return false;
                     }
-                    return property.enumNames[property.enumValueIndex].Contains(token);
+                    return property.enumNames[property.enumValueIndex].ToLower().Contains(token);
                 case SerializedPropertyType.Vector2:
-                    return property.vector2Value.ToString().Contains(token);
+                    return property.vector2Value.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Vector3:
-                    return property.vector3Value.ToString().Contains(token);
+                    return property.vector3Value.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Vector4:
-                    return property.vector4Value.ToString().Contains(token);
+                    return property.vector4Value.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Rect:
-                    return property.rectValue.ToString().Contains(token);
+                    return property.rectValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.ArraySize:
                     if (property.isArray)
                     {
-                        return property.arraySize.ToString().Contains(token);
+                        return property.arraySize.ToString().ToLower().Contains(token);
                     }
                     goto default;
                 case SerializedPropertyType.Character:
-                    return property.intValue.ToString().Contains(token);
+                    return property.intValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.AnimationCurve:
-                    return property.animationCurveValue.ToString().Contains(token);
+                    return property.animationCurveValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Bounds:
-                    return property.boundsValue.ToString().Contains(token);
+                    return property.boundsValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Quaternion:
-                    return property.quaternionValue.ToString().Contains(token);
+                    return property.quaternionValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.ExposedReference:
                     // ReSharper disable once Unity.NoNullPropagation
-                    return property.exposedReferenceValue?.name.Contains(token) ?? false;
+                    return property.exposedReferenceValue?.name.ToLower().Contains(token) ?? false;
                 case SerializedPropertyType.FixedBufferSize:
-                    return property.fixedBufferSize.ToString().Contains(token);
+                    return property.fixedBufferSize.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Vector2Int:
-                    return property.vector2IntValue.ToString().Contains(token);
+                    return property.vector2IntValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.Vector3Int:
-                    return property.vector3IntValue.ToString().Contains(token);
+                    return property.vector3IntValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.RectInt:
-                    return property.rectIntValue.ToString().Contains(token);
+                    return property.rectIntValue.ToString().ToLower().Contains(token);
                 case SerializedPropertyType.BoundsInt:
-                    return property.boundsIntValue.ToString().Contains(token);
+                    return property.boundsIntValue.ToString().ToLower().Contains(token);
 #if UNITY_2019_3_OR_NEWER
                 case SerializedPropertyType.ManagedReference:
                     // if (property.managedReferenceFullTypename.Contains(token))
@@ -477,13 +483,15 @@ namespace SaintsField.Editor.Utils
                         for (int i = 0; i < property.arraySize; i++)
                         {
                             if (SearchProp(property.GetArrayElementAtIndex(i), token))
+                            {
                                 return true;
+                            }
                         }
                         return false;
                     }
 
                     // ReSharper disable once LoopCanBeConvertedToQuery
-                    foreach (SerializedProperty child in SerializedUtils.GetPropertyChildren(property))
+                    foreach (SerializedProperty child in GetPropertyChildren(property))
                     {
                         if(SearchProp(child, token))
                         {